project/backend/requirements.txt
# Web Framework
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6

# Machine Learning
ultralytics>=8.0.0
torch>=2.0.0

# Computer Vision
opencv-python>=4.8.0
numpy>=1.24.0

# Database
pymongo>=4.5.0

# Environment Variables
python-dotenv>=1.0.0

# Data Validation (included with FastAPI but explicit for clarity)
pydantic>=2.0.0

